resolver wip

+++ resolver/internal/machine/machine.go
@@ -9,8 +9,8 @@ type Machine struct {
 }
 
 type Power struct {
-	Value int    `json:"value"`
-	Type  string `json:"type"`
+	Value float64 `json:"value"`
+	Type  string  `json:"type"`
 }
 
 const (
